requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

